+++ assets/test/debounce_test.js
@@ -10,7 +10,6 @@ let simulateInput = (input, val) => {
 
 let simulateKeyDown = (input, val) => {
   input.value = input.value + val;
-  DOM.dispatchEvent(input, "keydown")
   DOM.dispatchEvent(input, "input")
 }
 
@@ -55,6 +54,21 @@ describe("debounce", function() {
     DOM.dispatchEvent(el, "blur")
     expect(calls).toBe(1)
     expect(el.value).toBe("three")
+  })
+
+  test("triggers debounce on input blur caused by tab", async () => {
+    let calls = 0
+    let el = container().querySelector("input[name=debounce-200]")
+
+    el.addEventListener("input", e => {
+      DOM.debounce(el, e, "phx-debounce", 0, "phx-throttle", 0, () => calls++)
+    })
+    simulateInput(el, "one")
+    simulateInput(el, "two")
+    el.dispatchEvent(new KeyboardEvent("keydown", {bubbles: true, cancelable: true, key: "Tab"}))
+    DOM.dispatchEvent(el, "blur")
+    expect(calls).toBe(1)
+    expect(el.value).toBe("two")
   })
 
   test("triggers on timeout", done => {
@@ -105,7 +119,7 @@ describe("debounce", function() {
       simulateInput(el, "four")
       simulateInput(el, "five")
       simulateInput(el, "six")
-      after(600, () => {
+      after(1200, () => {
         expect(calls).toBe(1)
         expect(el.value).toBe("six")
         done()
@@ -113,34 +127,10 @@ describe("debounce", function() {
     })
   })
 
-
-  test("cancels trigger on phx-change", done => {
-    let calls = 0
-    let el = container().querySelector("input[name=debounce-200]")
-
-    el.addEventListener("input", e => {
-      DOM.debounce(el, e, "phx-debounce", 100, "phx-throttle", 200, () => calls++)
-    })
-    el.form.addEventListener("phx-change", () => {
-      el.value = "phx-changed"
-    })
-    simulateInput(el, "changed")
-    DOM.dispatchEvent(el.form, "phx-change")
-    after(100, () => {
-      expect(calls).toBe(0)
-      expect(el.value).toBe("phx-changed")
-      simulateInput(el, "changed again")
-      after(250, () => {
-        expect(calls).toBe(1)
-        expect(el.value).toBe("changed again")
-        done()
-      })
-    })
-  })
-
   test("cancels trigger on submit", done => {
     let calls = 0
-    let el = container().querySelector("input[name=debounce-200]")
+    let parent = container()
+    let el = parent.querySelector("input[name=debounce-200]")
 
     el.addEventListener("input", e => {
       DOM.debounce(el, e, "phx-debounce", 100, "phx-throttle", 200, () => calls++)
@@ -220,31 +210,6 @@ describe("throttle", function() {
         expect(el.innerText).toBe("now:1")
         done()
       })
-    })
-  })
-
-
-  test("cancels trigger on phx-change", done => {
-    let calls = 0
-    let el = container().querySelector("input[name=throttle-200]")
-    let otherInput = el.form.querySelector("input[name=debounce-200]")
-
-    el.addEventListener("input", e => {
-      DOM.debounce(el, e, "phx-debounce", 100, "phx-throttle", 200, () => calls++)
-    })
-    el.form.addEventListener("phx-change", () => {
-      el.value = "phx-changed"
-    })
-    simulateInput(el, "changed")
-    simulateInput(el, "changed2")
-    DOM.dispatchEvent(el.form, "phx-change", {triggeredBy: otherInput})
-    expect(calls).toBe(1)
-    expect(el.value).toBe("phx-changed")
-    simulateInput(el, "changed3")
-    after(100, () => {
-      expect(calls).toBe(2)
-      expect(el.value).toBe("changed3")
-      done()
     })
   })