+++ test/phoenix_live_view/diff_test.exs
@@ -4,6 +4,7 @@ defmodule Phoenix.LiveView.DiffTest do
   import Phoenix.LiveView.Helpers
 
   alias Phoenix.LiveView.{Socket, Diff, Rendered, Component}
+  alias Phoenix.LiveComponent.CID
 
   def basic_template(assigns) do
     ~L"""
@@ -594,7 +595,7 @@ defmodule Phoenix.LiveView.DiffTest do
       assert {MyComponent, "hello", _, _, _} = cid_to_component[1]
 
       assert_received {:mount, %Socket{endpoint: __MODULE__, assigns: assigns}}
-                      when assigns == %{flash: %{}, myself: 1}
+                      when assigns == %{flash: %{}, myself: %CID{cid: 1}}
 
       assert_received {:update, %{from: :component}, %Socket{assigns: %{hello: "world"}}}
       assert_received :render
@@ -615,7 +616,7 @@ defmodule Phoenix.LiveView.DiffTest do
       assert socket.fingerprints == {rendered.fingerprint, %{}}
 
       assert_received {:mount, %Socket{endpoint: __MODULE__, assigns: assigns}}
-                      when assigns == %{flash: %{}, myself: 1}
+                      when assigns == %{flash: %{}, myself: %CID{cid: 1}}
 
       assert_received :render
 
@@ -634,7 +635,7 @@ defmodule Phoenix.LiveView.DiffTest do
       assert socket.fingerprints != another_socket.fingerprints
 
       assert_received {:mount, %Socket{endpoint: __MODULE__, assigns: assigns}}
-                      when assigns == %{flash: %{}, myself: 2}
+                      when assigns == %{flash: %{}, myself: %CID{cid: 2}}
 
       assert_received :render
     end
@@ -665,7 +666,7 @@ defmodule Phoenix.LiveView.DiffTest do
       assert components == previous_components
 
       assert_received {:mount, %Socket{endpoint: __MODULE__, assigns: assigns}}
-                      when assigns == %{flash: %{}, myself: 1}
+                      when assigns == %{flash: %{}, myself: %CID{cid: 1}}
 
       assert_received {:update, %{from: :component}, %Socket{assigns: %{hello: "world"}}}
       assert_received :render
@@ -689,15 +690,15 @@ defmodule Phoenix.LiveView.DiffTest do
       assert components != previous_components
 
       assert_received {:mount, %Socket{endpoint: __MODULE__, assigns: assigns}}
-                      when assigns == %{flash: %{}, myself: 1}
+                      when assigns == %{flash: %{}, myself: %CID{cid: 1}}
 
       assert_received {:update, %{from: :component},
-                       %Socket{assigns: %{hello: "world", myself: 1}}}
+                       %Socket{assigns: %{hello: "world", myself: %CID{cid: 1}}}}
 
       assert_received :render
 
       assert_received {:update, %{from: :rerender},
-                       %Socket{assigns: %{hello: "world", myself: 1}}}
+                       %Socket{assigns: %{hello: "world", myself: %CID{cid: 1}}}}
 
       assert_received :render
       refute_received _